[Feat]: 전력량 24시간(1시간 단위), 7일간(1일 단위) 조회 구현
