+++ src/main/java/live/smoothing/sensordata/service/impl/KwhServiceImpl.java
@@ -2,8 +2,8 @@ package live.smoothing.sensordata.service.impl;
 
 import live.smoothing.sensordata.adapter.TopicAdapter;
 import live.smoothing.sensordata.dto.Kwh;
-import live.smoothing.sensordata.dto.PowerMetric;
-import live.smoothing.sensordata.dto.PowerMetricResponse;
+import live.smoothing.sensordata.dto.watt.PowerMetric;
+import live.smoothing.sensordata.dto.watt.PowerMetricResponse;
 import live.smoothing.sensordata.repository.KwhRepository;
 import live.smoothing.sensordata.service.KwhService;
 import lombok.RequiredArgsConstructor;
